split jet matching interface into headers and src files
