+++ GeneratorInterface/PartonShowerVeto/src/CombineMatchingInput.cc
@@ -1,0 +1,3 @@
+#include "GeneratorInterface/PartonShowerVeto/interface/CombineMatchingInput.h"
+
+using namespace Pythia8;
